refactor: close the cleanup plan — printf engine split + duplication remainder (#36, #37)
Item 37 (T3-6): extract the IEEE-754 printf engine out of
workspace/executor/builtins/text.{py,ts} into printf_format.{py,ts}.
text.py 776 -> 263 (-66%), text.ts 995 -> 266 (-73%). The move is
internal: no external importer reached the engine. printf keeps its own
escape tables because its grammar differs from echo's, matching how TS
already modelled it. Drops two orphans found in there: a duplicated doc
block on assignPrintfTarget and a trailing doc comment for handleRead,
which lives in vars.ts. New mirrored tests pin what the handle_printf
wrapper hides — the error list's order and count, and the
reuse-termination invariants — all GNU-pinned in debian:stable-slim.

The clause's second half, splitting execute_node, is deliberately left
out: every arm recurses through execute_fn and mutates shared
session/view state, so it needs its own review rather than this diff.

Item 36 (T3-4): add browser/src/resource/s3_alias.ts mirroring the node
twin and collapse the 14 browser alias configs onto it (793 -> 516
lines, net -158). The factory carries three knobs the survey actually
found: a derived endpoint rule, a region default (gcs and r2 report
auto), and forcePathStyle — which supabase alone pins true, the
difference a careless collapse would have flattened. Browser had no
equivalent of node's s3_aliases.test.ts; it does now, 47 tests over all
14.

Extract makeSearchBackedFind into core/generic/find.{py,ts} beside the
existing walkFind; chroma/dify find.py 139+139 -> 6+6 and find.ts
144+144 -> 21+21, with 13 new tests giving that logic its first direct
coverage. The python factory keeps the accessor as a TypeVar because
callable parameters are contravariant, matching the TS factory's <A>.

makeVectorSearch is not done, on evidence: the five search files are not
token-substitution-identical — qdrant and lancedb differ in config
vocabulary, score field and rendering — and mem0 has no TS twin, so a
five-backend shared factory cannot be mirrored.
