+++ typescript/packages/browser/src/resource/s3_alias.ts
@@ -1,0 +1,137 @@
+// ========= Copyright 2026 @ Strukto.AI All Rights Reserved. =========
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+// ========= Copyright 2026 @ Strukto.AI All Rights Reserved. =========
+
+import { redactConfigWithSchema, secretSchema, z } from '@struktoai/mirage-core/resource/secrets'
+import type { S3BrowserPresignedUrlProvider, S3Config } from './s3/config.ts'
+
+/**
+ * Shared shape for the browser's S3-compatible providers.
+ *
+ * Every alias is the same presigned-fetch config plus one endpoint rule,
+ * so the fields, the zod schema that drives redaction, and the conversion
+ * to S3Config live here once. A provider declares only its endpoint rule
+ * and whatever it genuinely adds. The node twin is
+ * `packages/node/src/resource/s3_alias.ts`; it carries credentials and a
+ * snake_case `normalize` where this one carries a
+ * `presignedUrlProvider`, which is one field's difference and not a
+ * reason to fork the family.
+ *
+ * Only `bucket` and the provider are required. `region` and `endpoint`
+ * are optional here — unlike node, where a regional provider must have a
+ * region — because a presigned URL already names its host, so an alias
+ * that can derive neither simply leaves both off the S3Config.
+ */
+export interface S3BrowserAliasConfig {
+  bucket: string
+  presignedUrlProvider: S3BrowserPresignedUrlProvider
+  region?: string
+  endpoint?: string
+  defaultContentType?: string
+  keyPrefix?: string
+}
+
+export const BROWSER_ALIAS_FIELDS = {
+  bucket: z.string(),
+  presignedUrlProvider: secretSchema(
+    z.custom<S3BrowserPresignedUrlProvider>((value) => typeof value === 'function'),
+  ),
+  region: z.string().optional(),
+  endpoint: z.string().optional(),
+  defaultContentType: z.string().optional(),
+  keyPrefix: z.string().optional(),
+}
+
+/**
+ * The shared schema, plus whatever one provider's endpoint rule needs
+ * (r2's `accountId`, oci's `namespace`, supabase's `projectRef`).
+ *
+ * The extra fields land after the shared ones rather than where each
+ * hand-written schema happened to put them. That is only visible in the
+ * key order of a redacted config, which nothing pins.
+ */
+export function browserAliasSchema<E extends z.ZodRawShape>(
+  extra: E,
+): z.ZodObject<typeof BROWSER_ALIAS_FIELDS & E> {
+  return z.object({ ...BROWSER_ALIAS_FIELDS, ...extra })
+}
+
+/**
+ * The common endpoint rule: one config field fills one URL template.
+ *
+ * Most providers key off `region`; r2 keys off `accountId` and supabase
+ * off `projectRef`. An absent or empty field yields no endpoint rather
+ * than a URL with a hole in it. oci needs two fields, and gcs and wasabi
+ * always resolve, so those three pass their own `endpointFor`.
+ */
+export function derivedEndpoint<C extends S3BrowserAliasConfig>(
+  key: keyof C & string,
+  template: (value: string) => string,
+): (config: C) => string | undefined {
+  return (config) => {
+    const value = config[key]
+    return typeof value === 'string' && value !== '' ? template(value) : undefined
+  }
+}
+
+export interface BrowserAlias<C, R, E extends string | undefined> {
+  resolvedEndpoint: (config: C) => string | E
+  toS3Config: (config: C) => S3Config
+  redact: (config: C) => R
+}
+
+/**
+ * Build one provider's config helpers from its endpoint rule.
+ *
+ * `endpointFor` is consulted only when the caller gave no explicit
+ * endpoint, and may answer `undefined` when the config lacks whatever
+ * the rule needs (a region, an account id); the S3Config then carries no
+ * endpoint at all. `regionDefault` materializes a region the provider
+ * always has (gcs and r2 report `auto`), and `forcePathStyle` is set
+ * only by supabase, whose gateway does not serve virtual-hosted style.
+ */
+export function makeBrowserS3Alias<
+  C extends S3BrowserAliasConfig,
+  R,
+  E extends string | undefined = string | undefined,
+>(options: {
+  schema: z.ZodObject<z.ZodRawShape>
+  endpointFor?: (config: C) => E
+  regionDefault?: string
+  forcePathStyle?: boolean
+}): BrowserAlias<C, R, E> {
+  const { schema, endpointFor, regionDefault, forcePathStyle } = options
+  const resolvedEndpoint = (config: C): string | E => {
+    if (config.endpoint !== undefined && config.endpoint !== '') return config.endpoint
+    return endpointFor === undefined ? (undefined as E) : endpointFor(config)
+  }
+  return {
+    resolvedEndpoint,
+    toS3Config: (config) => {
+      const endpoint = resolvedEndpoint(config)
+      const region = config.region ?? regionDefault
+      return {
+        bucket: config.bucket,
+        presignedUrlProvider: config.presignedUrlProvider,
+        ...(region !== undefined ? { region } : {}),
+        ...(endpoint !== undefined ? { endpoint } : {}),
+        ...(forcePathStyle !== undefined ? { forcePathStyle } : {}),
+        ...(config.defaultContentType !== undefined
+          ? { defaultContentType: config.defaultContentType }
+          : {}),
+        ...(config.keyPrefix !== undefined ? { keyPrefix: config.keyPrefix } : {}),
+      }
+    },
+    redact: (config) => redactConfigWithSchema(schema, config) as unknown as R,
+  }
+}
